+++ tests/Bridge/Doctrine/Common/Util/IdentifierManagerTraitTest.php
@@ -20,6 +20,8 @@ use ApiPlatform\Core\Metadata\Property\Factory\PropertyMetadataFactoryInterface;
 use ApiPlatform\Core\Metadata\Property\Factory\PropertyNameCollectionFactoryInterface;
 use ApiPlatform\Core\Metadata\Property\PropertyMetadata;
 use ApiPlatform\Core\Metadata\Property\PropertyNameCollection;
+use ApiPlatform\Core\Metadata\Resource\Factory\ResourceMetadataFactoryInterface;
+use ApiPlatform\Core\Metadata\Resource\ResourceMetadata;
 use ApiPlatform\Core\Tests\Fixtures\TestBundle\Document\Dummy as DummyDocument;
 use ApiPlatform\Core\Tests\Fixtures\TestBundle\Entity\Dummy;
 use ApiPlatform\Core\Tests\ProphecyTrait;
@@ -39,17 +41,18 @@ class IdentifierManagerTraitTest extends TestCase
 {
     use ProphecyTrait;
 
-    private function getIdentifierManagerTraitImpl(PropertyNameCollectionFactoryInterface $propertyNameCollectionFactory, PropertyMetadataFactoryInterface $propertyMetadataFactory)
+    private function getIdentifierManagerTraitImpl(PropertyNameCollectionFactoryInterface $propertyNameCollectionFactory, PropertyMetadataFactoryInterface $propertyMetadataFactory, ResourceMetadataFactoryInterface $resourceMetadataFactory)
     {
-        return new class($propertyNameCollectionFactory, $propertyMetadataFactory) {
+        return new class($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory) {
             use IdentifierManagerTrait {
                 IdentifierManagerTrait::normalizeIdentifiers as public;
             }
 
-            public function __construct(PropertyNameCollectionFactoryInterface $propertyNameCollectionFactory, PropertyMetadataFactoryInterface $propertyMetadataFactory)
+            public function __construct(PropertyNameCollectionFactoryInterface $propertyNameCollectionFactory, PropertyMetadataFactoryInterface $propertyMetadataFactory, ResourceMetadataFactoryInterface $resourceMetadataFactory)
             {
                 $this->propertyNameCollectionFactory = $propertyNameCollectionFactory;
                 $this->propertyMetadataFactory = $propertyMetadataFactory;
+                $this->resourceMetadataFactory = $resourceMetadataFactory;
             }
         };
     }
@@ -59,7 +62,7 @@ class IdentifierManagerTraitTest extends TestCase
      */
     public function testSingleIdentifier()
     {
-        [$propertyNameCollectionFactory, $propertyMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
+        [$propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
             'id',
         ]);
         $objectManager = $this->getEntityManager(Dummy::class, [
@@ -68,7 +71,7 @@ class IdentifierManagerTraitTest extends TestCase
             ],
         ]);
 
-        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory);
+        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory);
 
         $this->assertEquals($identifierManager->normalizeIdentifiers(1, $objectManager, Dummy::class), ['id' => 1]);
     }
@@ -79,7 +82,7 @@ class IdentifierManagerTraitTest extends TestCase
      */
     public function testSingleDocumentIdentifier()
     {
-        [$propertyNameCollectionFactory, $propertyMetadataFactory] = $this->getMetadataFactories(DummyDocument::class, [
+        [$propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory] = $this->getMetadataFactories(DummyDocument::class, [
             'id',
         ]);
         $objectManager = $this->getDocumentManager(DummyDocument::class, [
@@ -88,7 +91,7 @@ class IdentifierManagerTraitTest extends TestCase
             ],
         ]);
 
-        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory);
+        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory);
 
         $this->assertEquals($identifierManager->normalizeIdentifiers(1, $objectManager, DummyDocument::class), ['id' => 1]);
     }
@@ -98,7 +101,7 @@ class IdentifierManagerTraitTest extends TestCase
      */
     public function testCompositeIdentifier()
     {
-        [$propertyNameCollectionFactory, $propertyMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
+        [$propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
             'ida',
             'idb',
         ]);
@@ -111,7 +114,7 @@ class IdentifierManagerTraitTest extends TestCase
             ],
         ]);
 
-        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory);
+        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory);
 
         $this->assertEquals($identifierManager->normalizeIdentifiers('ida=1;idb=2', $objectManager, Dummy::class), ['ida' => 1, 'idb' => 2]);
     }
@@ -122,9 +125,9 @@ class IdentifierManagerTraitTest extends TestCase
     public function testInvalidIdentifier()
     {
         $this->expectException(PropertyNotFoundException::class);
-        $this->expectExceptionMessage('Invalid identifier "idbad=1;idb=2", "ida" was not found.');
+        $this->expectExceptionMessage('Invalid identifier "idbad=1;idb=2", "ida" was not found for resource "dummy".');
 
-        [$propertyNameCollectionFactory, $propertyMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
+        [$propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
             'ida',
             'idb',
         ]);
@@ -137,7 +140,7 @@ class IdentifierManagerTraitTest extends TestCase
             ],
         ]);
 
-        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory);
+        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory);
 
         $identifierManager->normalizeIdentifiers('idbad=1;idb=2', $objectManager, Dummy::class);
     }
@@ -149,6 +152,7 @@ class IdentifierManagerTraitTest extends TestCase
     {
         $propertyNameCollectionFactoryProphecy = $this->prophesize(PropertyNameCollectionFactoryInterface::class);
         $propertyMetadataFactoryProphecy = $this->prophesize(PropertyMetadataFactoryInterface::class);
+        $resourceMetadataFactoryProphecy = $this->prophesize(ResourceMetadataFactoryInterface::class);
 
         $nameCollection = ['foobar'];
 
@@ -164,8 +168,9 @@ class IdentifierManagerTraitTest extends TestCase
         $propertyMetadataFactoryProphecy->create($resourceClass, 'foobar')->willReturn(new PropertyMetadata());
 
         $propertyNameCollectionFactoryProphecy->create($resourceClass)->willReturn(new PropertyNameCollection($nameCollection));
+        $resourceMetadataFactoryProphecy->create($resourceClass)->willReturn(new ResourceMetadata('dummy'));
 
-        return [$propertyNameCollectionFactoryProphecy->reveal(), $propertyMetadataFactoryProphecy->reveal()];
+        return [$propertyNameCollectionFactoryProphecy->reveal(), $propertyMetadataFactoryProphecy->reveal(), $resourceMetadataFactoryProphecy->reveal()];
     }
 
     /**
@@ -218,9 +223,9 @@ class IdentifierManagerTraitTest extends TestCase
         DBALType::addType('uuid', UuidType::class);
 
         $this->expectException(InvalidIdentifierException::class);
-        $this->expectExceptionMessage('Invalid value "ida" provided for an identifier.');
+        $this->expectExceptionMessage('Invalid value "ida" provided for an identifier for resource "dummy".');
 
-        [$propertyNameCollectionFactory, $propertyMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
+        [$propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory] = $this->getMetadataFactories(Dummy::class, [
             'ida',
         ]);
         $objectManager = $this->getEntityManager(Dummy::class, [
@@ -229,7 +234,7 @@ class IdentifierManagerTraitTest extends TestCase
             ],
         ]);
 
-        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory);
+        $identifierManager = $this->getIdentifierManagerTraitImpl($propertyNameCollectionFactory, $propertyMetadataFactory, $resourceMetadataFactory);
 
         $identifierManager->normalizeIdentifiers('notanuuid', $objectManager, Dummy::class);
     }